determine runtime properly
In cases where the runtime is not installed in /kb/runtime builds that
use $(TPAGE) fail because of its dependency on the proper value of
DEPLOY_RUNTIME. This patch sets up a $(RUNTIME_BIN) that is computed
based first on the value of KB_RUNTIME and falling back to
DEPLOY_RUNTIME. It sets TPAGE accordingly.

I think this is safe; however for some reason the older definition of
TPAGE explicitly invoked perl on the tpage executable. I suspect this
may have been because of some improper installations of tpage (from the
perl Template::Toolkit library) but I am not sure.

Someone other than me should probably also test against htis.